Added todo app
I've added a To-do app, which is mostly complete. The main thing it currently lacks is the ability to rearrange items. It does support import/export, though, so the best way to rearrange items is to export the list (it will download in csv format), rearrange them there, and then re-import. Eventually I plan on adding support for tabs within the app to manage different projects with, and possibly color coding.